Create Famous places to visit Himachal Pradesh
Himachal is the one of the most famous tourist place of India.If you are on a lookout for fresh air and green mountains amidst the snow-covered ranges then these are the top destinations to visit throughout the year. Lots of people visit himachal to see the tradition, nature and for adventure, exploring and climbing the hills not only from India but also from the other countries.